# openapi-atm.yaml
openapi: 3.0.3
info:
  title: ATM System API
  version: 1.0.0
  description: |
    ATM 虚拟系统接口文档。
    用于课程作业中的 ATM 登录、查询余额、取款、存款、转账、修改密码、交易流水与凭条查询。

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: Auth
    description: 认证与会话接口
  - name: Account
    description: 账户信息接口
  - name: Transaction
    description: 交易接口
  - name: Receipt
    description: 凭条接口
  - name: Device
    description: ATM 设备接口

paths:
  /api/atm/auth/login:
    post:
      tags: [Auth]
      summary: 插卡登录
      description: 使用卡号和密码进行身份验证，成功后返回 ATM 会话信息。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: 密码错误或卡状态异常
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/atm/auth/logout:
    post:
      tags: [Auth]
      summary: 退卡退出
      description: 结束当前 ATM 会话。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SessionRequest'
      responses:
        '200':
          description: 退出成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/atm/auth/change-password:
    post:
      tags: [Auth]
      summary: 修改密码
      description: 修改当前银行卡登录密码。请求体主字段使用 sessionId，兼容旧字段 token；修改成功后当前会话失效，需要重新登录。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '200':
          description: 修改成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangePasswordResponse'
        '401':
          description: 原密码错误或会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/atm/accounts/balance:
    get:
      tags: [Account]
      summary: 查询余额
      description: 查询当前会话绑定账户的余额。
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
        '401':
          description: 会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/atm/accounts/profile:
    get:
      tags: [Account]
      summary: 查询账户信息
      description: 查询当前登录账户的基础信息。
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountProfileResponse'
        '401':
          description: 会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/atm/accounts/full-info:
    get:
      tags: [Account]
      summary: 查询完整账户信息
      description: 查询当前登录账户的完整信息，包含基础账户信息、脱敏手机号和账户状态。
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FullAccountProfileResponse'
        '401':
          description: 会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/atm/accounts/validate-transaction:
    get:
      tags: [Account]
      summary: 交易前身份和账户校验
      description: 在取款、存款、转账前校验当前 sessionId 对应的银行卡、账户和客户信息是否可用。
      parameters:
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 校验完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionValidationResponse'
        '401':
          description: 会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/atm/transactions/withdraw:
    post:
      tags: [Transaction]
      summary: 取款
      description: 发起取款交易，成功后返回交易编号和余额信息。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawRequest'
      responses:
        '200':
          description: 取款成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawResponse'
        '400':
          description: 金额无效、超出取款限额或余额不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 会话失效或未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: 银行卡或账户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/atm/transactions/deposit:
    post:
      tags: [Transaction]
      summary: 存款
      description: 发起存款交易。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositRequest'
      responses:
        '200':
          description: 存款成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositResponse'
        '400':
          description: 金额无效或超出存款限额
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 会话失效或未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: 银行卡或账户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/atm/transactions/transfer:
    post:
      tags: [Transaction]
      summary: 转账
      description: 发起转账交易。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferRequest'
      responses:
        '200':
          description: 转账成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: 金额无效、目标账户为空、自转账、超出转账限额或余额不足
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '401':
          description: 会话失效或未登录
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: 银行卡、源账户或目标账户不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/atm/transactions/{transactionId}:
    get:
      tags: [Transaction]
      summary: 查询交易详情
      description: 根据交易编号查询单笔交易详情。
      parameters:
        - $ref: '#/components/parameters/TransactionId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionDetailResponse'
        '400':
          description: 交易编号不能为空
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
        '404':
          description: 交易不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/atm/transactions/history:
    get:
      tags: [Transaction]
      summary: 查询交易流水
      description: 分页查询当前 sessionId 对应账户的交易流水。
      parameters:
        - $ref: '#/components/parameters/SessionId'
        - name: page
          in: query
          required: false
          schema:
            type: integer
            example: 1
        - name: size
          in: query
          required: false
          schema:
            type: integer
            example: 10
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionHistoryResponse'
        '401':
          description: 会话失效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/atm/receipts/{transactionId}:
    get:
      tags: [Receipt]
      summary: 查询交易凭条
      description: 根据交易编号和当前会话获取凭条信息。
      parameters:
        - $ref: '#/components/parameters/TransactionId'
        - $ref: '#/components/parameters/SessionId'
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptResponse'
        '404':
          description: 凭条不存在
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

  /api/atm/device/status:
    get:
      tags: [Device]
      summary: 查询 ATM 设备状态
      responses:
        '200':
          description: 查询成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeviceStatusResponse'

  /api/atm/device/cash-check:
    post:
      tags: [Device]
      summary: 检查 ATM 吐钞能力
      description: 用于在取款前判断设备现金是否充足；该接口只做检查，不扣减设备现金，真实扣减发生在取款成功时。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount:
                  type: number
                  format: double
                  example: 1000
      responses:
        '200':
          description: 检查完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CashCheckResponse'
        '400':
          description: 金额无效
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'

components:
  parameters:
    SessionId:
      name: sessionId
      in: query
      required: true
      description: ATM 会话 ID
      schema:
        type: string
        example: atm_session_xxx
    TransactionId:
      name: transactionId
      in: path
      required: true
      description: 交易编号
      schema:
        type: string
        example: TX202603010001

  schemas:
    ApiResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          nullable: true
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    LoginRequest:
      type: object
      required: [cardNo, password]
      properties:
        cardNo:
          type: string
          example: 6222020000000001
        password:
          type: string
          example: 123456

    SessionRequest:
      type: object
      required: [sessionId]
      properties:
        sessionId:
          type: string
          example: atm_session_xxx

    ChangePasswordRequest:
      type: object
      required: [sessionId, oldPassword, newPassword]
      properties:
        sessionId:
          type: string
          example: atm_session_xxx
        oldPassword:
          type: string
          example: 123456
        newPassword:
          type: string
          example: 654321

    ChangePasswordData:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 密码修改成功，请重新登录

    LoginData:
      type: object
      properties:
        sessionId:
          type: string
          example: atm_session_xxx
        accountId:
          type: integer
          format: int64
          example: 10001

    LoginResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 登录成功
        data:
          $ref: '#/components/schemas/LoginData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    ChangePasswordResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 密码修改成功，请重新登录
        data:
          $ref: '#/components/schemas/ChangePasswordData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    BalanceData:
      type: object
      properties:
        balance:
          type: number
          format: double
          example: 5000.00

    BalanceResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 查询成功
        data:
          $ref: '#/components/schemas/BalanceData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    AccountProfileData:
      type: object
      properties:
        customerName:
          type: string
          example: 张三
        cardNo:
          type: string
          example: 6222020000000001
        idCard:
          type: string
          example: 110101********1234
        accountNo:
          type: string
          example: ACC10001
        balance:
          type: number
          format: double
          example: 5000.00
        accountType:
          type: string
          example: 储蓄卡
        createTime:
          type: string
          format: date-time
          example: 2026-03-01T10:20:30

    FullAccountProfileData:
      allOf:
        - $ref: '#/components/schemas/AccountProfileData'
        - type: object
          properties:
            phone:
              type: string
              example: 138****0000
            status:
              type: string
              example: 正常

    AccountProfileResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 查询成功
        data:
          $ref: '#/components/schemas/AccountProfileData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    FullAccountProfileResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/FullAccountProfileData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    TransactionValidationData:
      type: object
      properties:
        valid:
          type: boolean
          example: true
        cardNo:
          type: string
          example: 6222020000000001
        accountId:
          type: integer
          format: int64
          example: 10001
        accountNo:
          type: string
          example: ACC10001
        balance:
          type: number
          format: double
          example: 5000.00
        customerName:
          type: string
          example: 张三
        message:
          type: string
          example: 验证通过
        accountType:
          type: integer
          example: 1
        status:
          type: string
          example: 正常

    TransactionValidationResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/TransactionValidationData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    WithdrawRequest:
      type: object
      required: [sessionId, amount, printReceipt]
      properties:
        sessionId:
          type: string
          example: atm_session_xxx
        amount:
          type: number
          format: double
          example: 1000
        printReceipt:
          type: boolean
          example: true

    DepositRequest:
      type: object
      required: [sessionId, amount, printReceipt]
      properties:
        sessionId:
          type: string
          example: atm_session_xxx
        amount:
          type: number
          format: double
          example: 2000
        printReceipt:
          type: boolean
          example: true

    TransferRequest:
      type: object
      required: [sessionId, targetAccountNo, amount, printReceipt]
      properties:
        sessionId:
          type: string
          example: atm_session_xxx
        targetAccountNo:
          type: string
          example: ACC20001
        amount:
          type: number
          format: double
          example: 800
        printReceipt:
          type: boolean
          example: true

    TransactionData:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        transactionId:
          type: string
          example: TX202603010001
        transactionType:
          type: integer
          description: 交易类型：1-取款，2-存款，3-转账
          example: 1
        transactionTypeName:
          type: string
          example: 取款
        amount:
          type: number
          format: double
          example: 1000
        balanceBefore:
          type: number
          format: double
          example: 5000
        balanceAfter:
          type: number
          format: double
          example: 4000
        transactionStatus:
          type: integer
          description: 交易状态：0-待处理，1-成功，2-失败，3-已撤销
          example: 1
        transactionStatusName:
          type: string
          example: 成功
        targetAccountNo:
          type: string
          nullable: true
          example: ACC20001
        failureReason:
          type: string
          nullable: true
          example: 余额不足
        description:
          type: string
          example: ATM取款
        createdAt:
          type: string
          format: date-time
          example: 2026-03-01T10:20:30
        completedAt:
          type: string
          format: date-time
          nullable: true
          example: 2026-03-01T10:20:31

    WithdrawData:
      type: object
      properties:
        transactionId:
          type: string
          example: TX202603010001
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 取款成功
        remainingBalance:
          type: number
          format: double
          example: 4000

    WithdrawResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/WithdrawData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    DepositData:
      type: object
      properties:
        transactionId:
          type: string
          example: TX202603010002
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 存款成功
        updatedBalance:
          type: number
          format: double
          example: 7000

    DepositResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/DepositData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    TransferData:
      type: object
      properties:
        transactionId:
          type: string
          example: TX202603010003
        success:
          type: boolean
          example: true
        message:
          type: string
          example: 转账成功
        remainingBalance:
          type: number
          format: double
          example: 4200

    TransferResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          $ref: '#/components/schemas/TransferData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    TransactionResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 交易成功
        data:
          $ref: '#/components/schemas/TransactionData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    TransactionDetailResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 查询成功
        data:
          $ref: '#/components/schemas/TransactionData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    TransactionHistoryItem:
      type: object
      properties:
        transactionId:
          type: string
          example: TX202603010001
        transactionType:
          type: string
          enum: [WITHDRAW, DEPOSIT, TRANSFER]
          example: WITHDRAW
        amount:
          type: number
          format: double
          example: 1000
        transactionStatus:
          type: string
          enum: [PENDING, SUCCESS, FAILED, CANCELLED]
          example: SUCCESS
        transactionTime:
          type: string
          format: date-time
          example: 2026-03-01T10:20:30

    TransactionHistoryData:
      type: object
      properties:
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 10
        total:
          type: integer
          example: 25
        records:
          type: array
          items:
            $ref: '#/components/schemas/TransactionHistoryItem'

    TransactionHistoryResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 查询成功
        data:
          $ref: '#/components/schemas/TransactionHistoryData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    ReceiptData:
      type: object
      properties:
        transactionId:
          type: string
          example: TX202603010001
        type:
          type: string
          enum: [WITHDRAW, DEPOSIT, TRANSFER]
          example: WITHDRAW
        amount:
          type: number
          format: double
          example: 1000
        balanceAfter:
          type: number
          format: double
          example: 4000
        time:
          type: string
          format: date-time
          example: 2026-03-01T10:20:30
        accountNo:
          type: string
          example: ACC10001

    ReceiptResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 获取成功
        data:
          $ref: '#/components/schemas/ReceiptData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    DeviceStatusData:
      type: object
      properties:
        atmCode:
          type: string
          example: ATM001
        location:
          type: string
          example: 一号教学楼大厅
        status:
          type: string
          enum: [RUNNING, MAINTENANCE, OUT_OF_SERVICE]
          example: RUNNING
        cashAvailable:
          type: number
          format: double
          example: 30000

    DeviceStatusResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 查询成功
        data:
          $ref: '#/components/schemas/DeviceStatusData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752

    CashCheckData:
      type: object
      properties:
        available:
          type: boolean
          example: true
        amount:
          type: number
          format: double
          example: 1000
        cashAvailable:
          type: number
          format: double
          example: 30000

    CashCheckResponse:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: 检查成功
        data:
          $ref: '#/components/schemas/CashCheckData'
        timestamp:
          type: integer
          format: int64
          example: 1778055092752