# transaction/src/main/java/com/atm/atmserver/dto/ReceiptResponse.java
package com.atm.atmserver.dto;

import lombok.Data;

import java.math.BigDecimal;
import java.time.LocalDateTime;

@Data
public class ReceiptResponse {
    private String transactionId;
    private String type;
    private BigDecimal amount;
    private BigDecimal balanceAfter;
    private LocalDateTime time;
    private String accountNo;
}