# transaction/src/main/java/com/atm/atmserver/dto/TransactionHistoryResponse.java
package com.atm.atmserver.dto;

import lombok.Data;

@Data
public class TransactionHistoryResponse {
    private Integer page;
    private Integer size;
    private Long total;
    private java.util.List<TransactionHistoryItem> records;
}